deeplearn-project/requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
importlib-resources==6.5.2
jinja2==3.1.6
kiwisolver==1.4.7
markupsafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing-extensions==4.15.0
zipp==3.23.0
